fixed warning levels that got broken; warning clean up

+++ src/engine/system/src/file/file.cc
@@ -59,8 +59,8 @@ ref<const File::Ticket> File::beginRead(u32 size, i64 offset, Allocator& arena) 
     u32 s;
     if (offset >= 0)
     {
-        be_assert(offset <= m_size, "reading past end of file");
-        be_assert(offset+size <= m_size, "reading past end of file");
+        be_assert((u64)offset <= m_size, "reading past end of file");
+        be_assert((u64)offset+size <= m_size, "reading past end of file");
         s = size ? size : be_checked_numcast<u32>(m_size - offset);
     }
     else
@@ -77,7 +77,7 @@ ref<const File::Ticket> File::beginRead(u32 size, i64 offset, Allocator& arena) 
 ref<const File::Ticket> File::beginWrite(const void* data, u32 size, i64 offset)
 {
     if (offset > 0)
-        be_assert(offset <= m_size, "writing past end of file");
+        be_assert((u64)offset <= m_size, "writing past end of file");
     else if (offset < 0)
         be_assert(offset+(i64)m_size+1 >= 0, "writing past end of file");
     ref<Ticket> t = ref<Ticket>::create(ticketPool(), byref(tempArena()), this, offset, size, data);
